requirements.txt
argparse==1.1
joblib==1.1
tiatoolbox==1.3.1
numpy>=1.23.5
opencv-python>=4.7.0.72
openslide-python>=1.2.0
torch>=1.13.1